+++ anvil/packages/schema/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./errors.js";
 export * from "./gameDescriptor.js";
+export * from "./contentSchemas.js";